Remove genversion and GENERATED_version in favor of linker overrides
The go:generate mechanism is well-suited for files which can be
deterministically generated, but version numbers inherently depend on
the environment.

Hence, let’s place the burden on the user to provide a suitable
version number. E.g. the Debian packaging could set the version number
to the Debian package version number.

An example invocation which uses the git version number is:

  go generate github.com/Debian/debiman/... && \
  go install \
    -ldflags "-X main.debimanVersion=$(git describe --always)" \
    github.com/Debian/debiman/...