idx2rwmap: do not sort by serving path
this is a revert of commit 4c4062620b581c2bc6adfa1a2ec9a488351baa76

I’m not sure what was non-deterministic about the output back then,
but redirect.IndexFromProto() returns entries in the order in which they were
loaded from the index. Maybe non-determinism was addressed at a different stage
in the meantime.

An unintended side-effect of the sorting was a preference for uppercase paths,
as they come first in ASCII.

Removing the sorting removes a behavioral difference between serving the
redirect index via auxserver and using idx2rwmap.

related to https://github.com/Debian/debiman/issues/161